backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

redis==5.0.1

python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

aiosmtplib==3.0.1
email-validator==2.1.0.post1

pillow==10.2.0
python-dateutil==2.8.2

aiofiles==23.2.1
httpx==0.26.0

pytest==7.4.4
pytest-asyncio==0.23.3